Consolidate tests
